requirements.txt
langchain>=0.2.14
langgraph>=0.2.21
openai>=1.40.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pandas>=2.2.2
python-dotenv>=1.0.1
neo4j>=5.22.0
rapidfuzz>=3.8.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
kaleido